✨: Add debugger vscode, column cinema entity, handle logic delete cineplex & cinema

+++ src/modules/cineplex/cineplex.service.ts
@@ -1,11 +1,19 @@
-import { Injectable, NotFoundException } from "@nestjs/common";
+import {
+  BadRequestException,
+  forwardRef,
+  Inject,
+  Injectable,
+  NotFoundException,
+} from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from "typeorm";
 import { Transactional } from "typeorm-transactional-cls-hooked";
 
 import type { PageDto } from "../../common/dto/page.dto";
+import { CinemaEntity } from "../../entity/cinema.entity";
 import { CineplexEntity } from "./../../entity/cineplex.entity";
 import { ValidatorService } from "./../../shared/services/validator.service";
+import { CinemaService } from "./../cinema/cinema.service";
 import type { CineplexDto } from "./dto/cineplex.dto";
 import type { CineplexesPageOptionsDto } from "./dto/cineplex-page-option.dto";
 import { CreateCineplexDto } from "./dto/create-cineplex.dto";
@@ -16,6 +24,10 @@ export class CineplexService {
   constructor(
     @InjectRepository(CineplexEntity)
     private cineplexRepository: Repository<CineplexEntity>,
+    @InjectRepository(CinemaEntity)
+    private cinemaRepository: Repository<CinemaEntity>,
+    @Inject(forwardRef(() => CinemaService))
+    private cinemaService: CinemaService,
     private validatorService: ValidatorService
   ) {}
 
@@ -67,6 +79,36 @@ export class CineplexService {
   async deleteCineplex(cineplexId: Uuid): Promise<CineplexEntity> {
     const cineplexEntity = await this.getCineplex(cineplexId);
 
+    const preUpdateCinemas = await this.preUpdateCinemaByCineplexId(
+      cineplexEntity.id
+    );
+
+    if (preUpdateCinemas.length === 0) {
+      throw new BadRequestException();
+    }
+
     return this.cineplexRepository.remove(cineplexEntity);
+  }
+
+  @Transactional()
+  async preUpdateCinemaByCineplexId(cineplexId: Uuid): Promise<CinemaEntity[]> {
+    const cinemaEntity = await this.cinemaRepository.find({
+      where: {
+        cineplex: {
+          id: cineplexId,
+        } as never,
+      },
+    });
+
+    if (cinemaEntity.length === 0) {
+      throw new NotFoundException();
+    }
+
+    const cinemas = cinemaEntity.map((cinema) => ({
+      ...cinema,
+      cineplex: null as never,
+    }));
+
+    return this.cinemaRepository.save(cinemas);
   }
 }
